Moved files to view
