+++ app/src/main/java/com/example/kemptonparkaog/View/BaseFragment.kt
@@ -1,4 +1,4 @@
-package com.example.kemptonparkaog
+package com.example.kemptonparkaog.View
 
 import android.content.Context
 import androidx.annotation.LayoutRes
